+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v2.0.3 - 2022-05-08
+## v2.0.4 - 2022-05-11
 
 ### Changed
 
@@ -24,6 +24,7 @@
 
 - `Scru128Id#counter_hi()`, `Scru128Id#counter_lo()`, `Scru128Id#entropy()`
 - `Scru128Id#to_u128()`, `Scru128Id#to_bytes()`
+- Compact binary serde format
 
 ## v1.0.0 - 2022-01-03